+++ include/exiv2/value.hpp
@@ -98,7 +98,7 @@ namespace Exiv2 {
           @param byteOrder Applicable byte order (little or big endian).
           @return Number of bytes written.
         */
-        virtual long copy(byte* buf, ByteOrder byteOrder) const =0;
+        virtual size_t copy(byte* buf, ByteOrder byteOrder) const =0;
         //! Return the number of components of the value
         virtual size_t count() const =0;
         //! Return the size of the value in bytes
@@ -244,9 +244,7 @@ namespace Exiv2 {
 
         explicit DataValue(TypeId typeId =undefined);
 
-        DataValue(const byte* buf,
-                  long len, ByteOrder byteOrder =invalidByteOrder,
-                  TypeId typeId =undefined);
+        DataValue(const byte* buf, size_t len, ByteOrder byteOrder =invalidByteOrder, TypeId typeId =undefined);
 
         ~DataValue() override = default;
 
@@ -273,7 +271,7 @@ namespace Exiv2 {
           @param byteOrder Byte order. Not needed.
           @return Number of characters written.
         */
-        long copy(byte* buf, ByteOrder byteOrder = invalidByteOrder) const override;
+        size_t copy(byte* buf, ByteOrder byteOrder = invalidByteOrder) const override;
         size_t count() const override;
         size_t size() const override;
         std::ostream& write(std::ostream& os) const override;
@@ -346,7 +344,7 @@ namespace Exiv2 {
           @param byteOrder Byte order. Not used.
           @return Number of characters written.
         */
-        long copy(byte* buf, ByteOrder byteOrder = invalidByteOrder) const override;
+        size_t copy(byte* buf, ByteOrder byteOrder = invalidByteOrder) const override;
         size_t count() const override;
         size_t size() const override;
         int64_t toInt64(size_t n = 0) const override;
@@ -531,7 +529,7 @@ namespace Exiv2 {
         //! @name Accessors
         //@{
         UniquePtr clone() const { return UniquePtr(clone_()); }
-        long copy(byte* buf, ByteOrder byteOrder) const override;
+        size_t copy(byte* buf, ByteOrder byteOrder) const override;
         /*!
           @brief Write the comment in a format which can be read by
           read(const std::string& comment).
@@ -614,7 +612,7 @@ namespace Exiv2 {
           @param byteOrder Byte order. Not used.
           @return Number of characters written.
         */
-        long copy(byte* buf, ByteOrder byteOrder = invalidByteOrder) const override;
+        size_t copy(byte* buf, ByteOrder byteOrder = invalidByteOrder) const override;
         //@}
 
         //! @name Manipulators
@@ -974,7 +972,7 @@ namespace Exiv2 {
           @param byteOrder Byte order. Not used.
           @return Number of characters written.
         */
-        long copy(byte* buf, ByteOrder byteOrder = invalidByteOrder) const override;
+        size_t copy(byte* buf, ByteOrder byteOrder = invalidByteOrder) const override;
 
         //! Return date struct containing date information
         virtual const Date& getDate() const;
@@ -1072,7 +1070,7 @@ namespace Exiv2 {
           @param byteOrder Byte order. Not used.
           @return Number of characters written.
         */
-        long copy(byte* buf, ByteOrder byteOrder = invalidByteOrder) const override;
+        size_t copy(byte* buf, ByteOrder byteOrder = invalidByteOrder) const override;
         //! Return time struct containing time information
         virtual const Time& getTime() const;
         size_t count() const override;
@@ -1142,7 +1140,7 @@ namespace Exiv2 {
         // The default c'tor and this one can be combined, but that causes MSVC 7.1 to fall on its nose
         explicit ValueType(TypeId typeId);
         //! Constructor.
-        ValueType(const byte* buf, long len, ByteOrder byteOrder, TypeId typeId =getType<T>());
+        ValueType(const byte* buf, size_t len, ByteOrder byteOrder, TypeId typeId =getType<T>());
         //! Constructor.
         explicit ValueType(const T& val, TypeId typeId =getType<T>());
         //! Copy constructor
@@ -1173,7 +1171,7 @@ namespace Exiv2 {
         //! @name Accessors
         //@{
         UniquePtr clone() const { return UniquePtr(clone_()); }
-        long copy(byte* buf, ByteOrder byteOrder) const override;
+        size_t copy(byte* buf, ByteOrder byteOrder) const override;
         size_t count() const override;
         size_t size() const override;
         std::ostream& write(std::ostream& os) const override;
@@ -1457,7 +1455,7 @@ namespace Exiv2 {
     }
 
     template <typename T>
-    ValueType<T>::ValueType(const byte* buf, long len, ByteOrder byteOrder, TypeId typeId) : Value(typeId)
+    ValueType<T>::ValueType(const byte* buf, size_t len, ByteOrder byteOrder, TypeId typeId) : Value(typeId)
     {
         read(buf, len, byteOrder);
     }
@@ -1535,9 +1533,9 @@ namespace Exiv2 {
     }
 
     template<typename T>
-    long ValueType<T>::copy(byte* buf, ByteOrder byteOrder) const
+    size_t ValueType<T>::copy(byte* buf, ByteOrder byteOrder) const
     {
-        long offset = 0;
+        size_t offset = 0;
         for (auto i = value_.begin(); i != value_.end(); ++i) {
             offset += toData(buf + offset, *i, byteOrder);
         }
@@ -1553,7 +1551,7 @@ namespace Exiv2 {
     template<typename T>
     size_t ValueType<T>::size() const
     {
-        return static_cast<long>(TypeInfo::typeSize(typeId()) * value_.size());
+        return TypeInfo::typeSize(typeId()) * value_.size();
     }
 
     template<typename T>
